Babel preset: add support for Flow target, transpile __DEV__ expression
This change allows us to set target of our Babel transpilation. We can either set JS (default) or Flow. This is necessary because we need to do transpilations for both environments (and in future possibly any other like TS for example). Transpilation for JS applies the standard rules we used until today. Flow transpilation adds declaration of  `__DEV__` expression so external projects are going to be without any Flow issues. This pattern can be later used even for other expressions like for example `suppress_type=$FlowFixMe` which causes the same issues in external projects.

This is first part of the changes. Changes in NPM publisher will follow.

kiwicom-source-id: 210252f7c78698a52741cea4b0e8a748c57f99fe