Babel preset: fix 'invariant' transpilation in production env
I don't understand 100% what is going on but seems like env preset transpiles correctly on the first function call and ignores the second one. I don't think this is the right fix but I wasn't able to get my head around it so I just cloned the first transpiled version and adjusted it. Some help would be appreciated.

Closes: https://gitlab.skypicker.com/incubator/universe/issues/477

kiwicom-source-id: 6612ad9682cf257aece9a66ddd5ba9c42d047e4b