Babel preset: add documentation for __DEV__ expression
Some people try to use this expression like a global variable (`__DEV__ = false`) but it results in invalid code since it's not a variable. I don't know how to prevent this and I am not even sure we should (because it's like writing `true = false`). Explanation in the docs should help a little bit.

Closes: https://gitlab.skypicker.com/incubator/universe/issues/485

kiwicom-source-id: 73506833dea628ace48f02871b6c272b75064b37