Add filePath to dev-expression-check.js
I spent some time trying to find out why I got this error.
It seems I was trying to transpile already transpiled code.
Printing the file path should make debugging this easier.
As for the test, it is possible to mock the path, adding filename to
babel.transform second parameter, but it will use absolute path,
so the snapshot will always fail.

kiwicom-source-id: b8ca629486b6b3d0fab35f0dec3f5691f84d99b4